requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
faiss-cpu==1.8.0
sentence-transformers==3.0.1
transformers==4.43.3
numpy==1.26.4
scikit-learn==1.5.1
google-generativeai==0.7.2
pypdf==4.3.1